+++ sdk/storage/storage-file-datalake/src/clients.ts
@@ -329,6 +329,23 @@ export class DataLakePathClient extends StorageClient {
     const { span, updatedOptions } = createSpan("DataLakePathClient-create", options);
     try {
       ensureCpkIfSpecified(options.customerProvidedKey, this.isHttps);
+      let expiryOptions: FileExpiryMode | undefined;
+      let expiresOn: string | undefined;
+      if (options.expiresOn) {
+        if (Number.isFinite(options.expiresOn)) {
+          expiryOptions = "RelativeToNow";
+          expiresOn = Math.round(options.expiresOn! as number).toString();
+        } else {
+          const fileExpiryTime = options.expiresOn as Date;
+          if (!fileExpiryTime) {
+            throw new Error(`Value for expiresOn is invalid.`);
+          } else {
+            expiryOptions = "Absolute";
+            expiresOn = fileExpiryTime.toUTCString();
+          }
+        }
+      }
+
       return await this.pathContext.create({
         ...options,
         resource: resourceType,
@@ -336,6 +353,9 @@ export class DataLakePathClient extends StorageClient {
         modifiedAccessConditions: options.conditions,
         properties: toProperties(options.metadata),
         cpkInfo: options.customerProvidedKey,
+        acl: options.acl ? toAclString(options.acl) : undefined,
+        expiryOptions: expiryOptions,
+        expiresOn: expiresOn,
         ...convertTracingToRequestOptionsBase(updatedOptions),
       });
     } catch (e: any) {