+++ sdk/storage/storage-file-datalake/src/utils/constants.ts
@@ -2,7 +2,7 @@
 // Licensed under the MIT license.
 
 export const SDK_VERSION: string = "12.9.1";
-export const SERVICE_VERSION: string = "2021-06-08";
+export const SERVICE_VERSION: string = "2021-08-06";
 
 export const KB: number = 1024;
 export const MB: number = KB * 1024;